fix npe

+++ server/src/main/java/org/apache/iotdb/db/metadata/mtree/store/disk/file/MockSchemaFile.java
@@ -109,6 +109,9 @@ public class MockSchemaFile implements ISchemaFile {
   @Override
   public void deleteMNode(IMNode targetNode) {
     IMNode removedNode = getSegment(targetNode.getParent()).remove(targetNode.getName());
+    if (removedNode == null || removedNode.isMeasurement()) {
+      return;
+    }
     deleteMNodeRecursively(removedNode);
   }
